+++ src/scripts/rule-tests/cases/exec-sql-exec-immediate-create-table-singleline.ts
@@ -1,28 +1,24 @@
 import type { Case } from '../types';
 
 const testCase: Case = {
-  name: 'exec-sql-exec-immediate-multiline',
+  name: 'exec-sql-exec-immediate-create-table-singleline',
   input: [
     '**free',
     'exec sql',
     'begin',
-    '  execute immediate',
-    '    \'create table qtemp/demo_ins (',
-    '       ID char(10),',
-    '       NAME char(10)',
-    '     )\';',
+    '  execute immediate \'create table qtemp/demo (ID char(10), NAME char(10))\';',
     'end;',
     'end-exec;'
   ].join('\n'),
   mustInclude: [
     'execute immediate',
-    '\'create table qtemp/demo_ins (',
+    '\'create table qtemp/demo (',
     'ID    char(10),',
     'NAME  char(10)',
     ')\';'
   ],
   mustExclude: [
-    "+ '"
+    '(ID char(10), NAME char(10))'
   ]
 };